migration fix

+++ var/migrations/1511213524_2_address.up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE countries(
-  id INTEGER PRIMARY KEY ,
+  id SMALLSERIAL PRIMARY KEY ,
   name VARCHAR
 );
